requirements (1).txt
discord.py==2.3.2
flask
python-dotenv
requests
tlgbotfwk